+++ CHANGES.md
@@ -3,6 +3,7 @@
 * clean up `get_release_asset_redirect()`
 * add geopattern-icon default plugin icon
 * use direct GitHub release asset URL, not redirect URL
+* add filter `gu_plugin_assets_dir` to specify location of repository banners/icons
 
 #### 10.7.0 / 2022-03-06
 * fix long standing object cache conflict with refresh cache and missing GitHub subtab by always showing GitHub subtab